+++ src/features/dreamsAdEngine/types/interfaces.ts
@@ -95,5 +95,7 @@ declare global {
     dreamsSlotsToUpdate: Slot[];
     dreamsAllSlots: Slot[];
     apstag: any;
+    pbjs?: any;
+    dataLayer?: any[];
   }
 }